Added Jump Map Subsector Generation
Adds Subsector generation as a Jump Map. The option rolls up systems until it gets one with an A-class starport. This is designated as the capital while the sector out to a certain distance is mapped. If the range is further than the "known" radius, it simply notes that there is an unexplored system in that direction.